% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example1}
\alias{example1}
\title{Simulated estimation results}
\format{workspace}
\source{
Simulation
}
\description{
This object contains simulated data according to a continuous Rating Scale
Model. The data set contains 6 items and 400 observations.
}
\keyword{datasets}

