% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example3.r
\docType{data}
\name{example3}
\alias{example3}
\title{Simulated data set for estimating a Rasch model}
\format{workspace}
\source{
Simulation
}
\description{
This object contains simulated data according to a
dichotomous Rasch Model. The data set contains 10 items and 500 observations.
}
\keyword{data}

