% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{data}
\name{exdat}
\alias{exdat}
\title{Data set exdat}
\format{workspace}
\source{
simulation
}
\description{
This object contains simulated data that provide as example data for the examples of the function. The data set consists of 15 items and 500 persons
#'
}
\keyword{datasets}

