% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{data}
\name{extraversion}
\alias{extraversion}
\title{Data set B5PO-R personality questionnaire}
\format{workspace}
\source{
Study
}
\description{
This object contains data from. The items are from the extraversion scale.
#'
}
\references{
Siptroth, V. and Kubinger, K. D.
}
\keyword{datasets}

