% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{data}
\name{reason.test}
\alias{reason.test}
\title{Data set META reasoning test.}
\format{A matrix with 22 variables and 380 observations. Variables 'I1' to 'I11' contain the responses to the eleven items, 'BT1' to 'BT11' the response times for each item in seconds.}
\source{
Study
}
\description{
This object contains data from the reasoning test 'META' by Gatternig and Kubinger (1994). The test includes 11 encoding tasks.
}
\references{
Gatternig, J. and Kubinger, K. D. (1994). Erkennen von Metaregeln. Frankfurt: Swets.
}
\keyword{datasets}
