\name{example1}
\docType{data}
\alias{example1}
\title{Simulated estimation results}
\description{
  This object contains simulated data according to a continuous Rating Scale Model. The data set contains 6 items and 400 observations.
}
\usage{
data(example1)}
\format{workspace}
\source{Simulation}

\keyword{datasets}