% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{example2}
\alias{example2}
\title{Simulated data set}
\format{workspace}
\source{
Simulation
}
\description{
This object contains simulated data according to a constrained
multidimensional polytomous Rasch model. The data set contains 8 items and
500 observations.
}
\keyword{datasets}

