\name{example2}
\docType{data}
\alias{example2}
\title{Simulated data set}
\description{
  This object contains simulated data according to a constrained multidimensional polytomous Rasch model. The data set contains 8 items and 500 observations.
}
\usage{
data(example2)}
\format{workspace}
\source{Simulation}

\keyword{ datasets }