% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{data}
\name{extraversion}
\alias{extraversion}
\title{Data set extraversion}
\format{A matrix with 8 variables and 150 observations.}
\source{
Study
}
\description{
This object contains data from an extraversion scale . The data set consists of 8 items and 150 persons.
}
\keyword{datasets}
