% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{data}
\name{extraversion}
\alias{extraversion}
\title{Data set B5PO-R personality questionnaire}
\format{workspace}
\source{
Study
}
\description{
This object contains data from the B5PO-R (Kubinger and Siptroth, in prep.), the revised version of the B5PO (Holocher-Ertl, Kubinger and Menghin, 2003). The items are from the extraversion scale.
#'
}
\references{
Holocher-Ertl, S., Kubinger, K. D. and Menghin, S. (2003). Big Five Plus One Persoenlichkeitsinventar (B5PO). Moedling: Wiener Testsystem/Schuhfried.
}
\keyword{datasets}

