% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcIRT-package.R
\docType{package}
\name{pcIRT-package}
\alias{pcIRT-package}
\alias{pcIRT}
\title{IRT Models for Polytomous and Continuous Item Responses}
\description{
The multidimensional polytomous Rasch model (Rasch,
1961) can be estimated with pcIRT. It provides functions to set linear restrictions on the item category
parameters of this models. With this functions it is possible to test
whether item categories can be collapsed or set as linear dependent. Thus it
is also possible to test whether the multidimensional model can be reduced
to a unidimensional model that is whether item categories represent a
unidimensional continuum. For this case the scoring parameter of the
categories is estimated.
}
\details{
\tabular{ll}{ Package: \tab pcIRT\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2013-11-13\cr License: \tab GPL-3\cr }
}
\examples{

#simulate data set according to the multidimensional polytomous Rasch model (MPRM)
simdat <- simMPRM(rbind(matrix(c(-1.5,0.5,0.5,1,0.8,-0.3, 0.2,-1.2), ncol=4),0), 500)

#estimate MPRM item parameters
res_mprm <- MPRM(simdat$datmat)

summary(res_mprm)


}
\references{
Andersen, E. B. (1995). Polytomous Rasch models and their
estimation. In G. H. Fischer and I. Molenaar (Eds.). Rasch Models -
Foundations, Recent Developements, and Applications. Springer.

Fischer, G. H. (1974). Einfuehrung in die Theorie psychologischer Tests
[Introduction to test theory]. Bern: Huber.

Hohensinn, C. (2018). pcIRT: An R Package for Polytomous and Continuous Rasch Models. 
Journal of Statistical Software, Code Snippets, 84(2), 1-14. doi:10.18637/jss.v084.c02

Mueller, H. (1987). A Rasch model for continuous ratings. Psychometrika, 52,
165-181.

Rasch, G. (1961). On general laws and the meaning of measurement in
psychology, Proceedings Fourth Berekely Symposium on Mathematical
Statistiscs and Probability 5, 321-333.
}
\seealso{
\code{\link{MPRM}} \code{\link{CRSM}}
}
\author{
Christine Hohensinn Maintainer: Christine Hohensinn
<research@christinehohensinn.at>
}
\keyword{Continuous}
\keyword{IRT}
\keyword{Item}
\keyword{Model}
\keyword{Rasch}
\keyword{Rating}
\keyword{Response}
\keyword{Scale}
\keyword{Theory}
\keyword{model}
\keyword{multidimensional}
\keyword{package}
\keyword{polytomous}
\keyword{psychometrics}
